[Change] Simpler map visibility (client,server,shared lib)

+++ server/fight/LocalClientHandlerFight.cpp
@@ -256,7 +256,7 @@ bool Client::botFightCollision(CommonMap *map,const COORD_TYPE &x,const COORD_TY
     std::pair<uint8_t,uint8_t> pos(x,y);
     if(static_cast<MapServer *>(map)->botsFightTrigger.find(pos)!=static_cast<MapServer *>(map)->botsFightTrigger.cend())
     {
-        const std::vector<uint16_t> &botList=static_cast<MapServer *>(map)->botsFightTrigger.at(pos);
+        const std::vector<uint8_t> &botList=static_cast<MapServer *>(map)->botsFightTrigger.at(pos);
         unsigned int index=0;
         while(index<botList.size())
         {
@@ -278,7 +278,7 @@ bool Client::botFightCollision(CommonMap *map,const COORD_TYPE &x,const COORD_TY
     return false;
 }
 
-bool Client::botFightStart(const uint16_t &botFightId)
+bool Client::botFightStart(const uint8_t &botFightId)
 {
     if(isInFight())
     {
